refactor: move Canton config to a separate module
